% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_analysis.R
\name{dirichlet_multinomial}
\alias{dirichlet_multinomial}
\title{Title}
\usage{
dirichlet_multinomial(data.obj, components = 3, level = "Genus")
}
\arguments{
\item{data.obj}{A stromboli data object}

\item{components}{The number of dirichlet components / community types}

\item{level}{Taxonomic level of interest}
}
\value{
A list of plots showing the taxonomic drivers of each component
}
\description{
Title
}
