% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{generate_ordination2}
\alias{generate_ordination2}
\title{Title}
\usage{
generate_ordination2(data.obj, dist.obj, dist.names = c("UniFrac",
  "GUniFrac", "WUniFrac", "BC"), grp.name, adj.name = NULL,
  emp.lev = NULL, strata = NULL, pc.separate, pca.method = "cmd",
  ann = NULL, sub = NULL, clab = 1, cex.pt = 1.25, ellipse = T,
  cstar = 1, wid = 900, hei = 600, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
