% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{generate_rarefy_curve2}
\alias{generate_rarefy_curve2}
\title{Generate rarefaction curve}
\usage{
generate_rarefy_curve2(data.obj, grp.name, measures = c("Observed",
  "Chao1", "Shannon", "InvSimpson"), depth = NULL, iter.no = 5,
  npoint = 10, seed = 123,
  gg.cmd = "theme(legend.justification=c(1,0), legend.position=c(1,0))",
  wid = 5, hei = 5, error = "se")
}
\arguments{
\item{data.obj}{A data object created by load_data()}

\item{grp.name}{The variable of interest}

\item{measures}{A list of diversity measures to use (Default includes "Shannon", "InvSimpson", "Chao1", and "Observed")}

\item{depth}{Rarefaction depth}

\item{iter.no}{Number of iterations (Optional, default: 5)}

\item{npoint}{Number of points to plot (Optional, default: 10)}

\item{seed}{Random seed to use}

\item{gg.cmd}{Additional ggplot commands to use (Optional)}

\item{error}{Type of error, either "se" or "ci" (Default: "se")}
}
\value{
Ggplot object containing rarefaction curve for all specified diversity measures
}
\description{
Generate rarefaction curve
}
\examples{
data("Constipation")
generate_rarefy_curve2(data.obj, grp.name="Visit", measures = c("Observed", "Chao1", "Shannon", "InvSimpson"), depth=10000, iter.no=5)
}
