% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_stacked_barplots}
\alias{generate_stacked_barplots}
\title{Generate stacked barplots}
\usage{
generate_stacked_barplots(data.obj, taxa.level, grp.name)
}
\arguments{
\item{data.obj}{A data object created by load_data()}

\item{taxa.level}{The taxonomic level to use in the plot (e.g., "Phylum", "Family", etc. )}

\item{grp.name}{The variable of interest}
}
\value{
A list of two barplots: a per-sample barplot and an aggregated barplot
}
\description{
Generate stacked barplots
}
\examples{
data(Constipation)
barplots <- generate_stacked_barplots(data.obj, taxa.level = 'Phylum', grp.name='Visit')
barplots$aggregate
barplots$by_sample
}
