% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{generate_summary_string}
\alias{generate_summary_string}
\title{Generate summary string}
\usage{
generate_summary_string(data.obj)
}
\arguments{
\item{data.obj}{A data object created by load_data}
}
\value{
A character string containing some basic sequencing statistics: the number of samples, median reads per sample, range of reads per sample, number of OTUs/ASVs
}
\description{
Generate summary string
}
