% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobj_functions.R
\name{load_data}
\alias{load_data}
\title{Load microbiome data files and create data object}
\usage{
load_data(otu.file, map.file, tree.file = NULL, load.map = TRUE,
  parseFunction = parse_taxonomy_default, version = "Old",
  species = TRUE, filter.no = 1, rff = FALSE, dep = NULL,
  norm = "TSS", level = "OTU", intersect.no = 4, winsor = FALSE,
  winsor.qt = 0.97, ko.file = NULL, cog.file = NULL,
  ko.ann.file = NULL, meta.sep = "\\t", quote = "\\"",
  comment = "", read.gg = FALSE, seed = 1234, ...)
}
\arguments{
\item{otu.file}{A taxonomy-annotated BIOM file (either HDF5 or old format is fine)}

\item{map.file}{Tab-delimited metadata file, one row per sample.}

\item{tree.file}{A newick-formatted tree file}

\item{load.map}{Load mapping file or no? Default is TRUE}

\item{parseFunction}{Either parse_taxonomy_greengenes or parse_taxonomy_default. Default is parse_taxonomy_greengenes.}

\item{version}{Either "New" for HDF5 biom files or "Old" for old biom files. Default is "Old".}

\item{species}{Include species annotations, either TRUE or FALSE. Default is TRUE.}

\item{filter.no}{Filter out OTUs with with read support below this number. Default is 1 (filters out singletons).}

\item{rff}{Perform rarefaction, TRUE or FALSE. Default is FALSE.}

\item{dep}{Rarefaction depth, set to any integer when rff is TRUE.}

\item{norm}{Normalization method. Default is "TSS". "GMPR" is also supported.}

\item{level}{Normalization level. Default is "OTU".}

\item{intersect.no}{Argument for when "GMPR" normalization is selected. The minimum number of shared features between sample pair, where the ratio is calculated. Default value is 4.}

\item{winsor}{Perform winsorization, TRUE or FALSE. Default is TRUE.}

\item{winsor.qt}{Winsorization quartile to use. Default is 0.97.}

\item{ko.file}{KO biom file generated by Picrust1 (optional).}

\item{cog.file}{COG biom file generated by Picrust1 (optional).}

\item{ko.ann.file}{KO annotation file (optional).}

\item{meta.sep}{Delimiter in mapping file. Default is tab character.}

\item{quote}{Quote character}

\item{comment}{Indicates any character marking a comment line in the metadata file. Default assumption is that there are no comments}

\item{read.gg}{Set to TRUE if reading in a GreenGenes tree. Default is FALSE.}

\item{seed}{RNG seed. Default is 1234.}

\item{...}{Any additional parameters to pass to load_data()}
}
\value{
Returns a data object, tying together the information from the metadata, biom file, and the tree.
}
\description{
Load microbiome data files and create data object
}
