% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{perform_betadisper_test}
\alias{perform_betadisper_test}
\title{Title}
\usage{
perform_betadisper_test(data.obj, dist.obj, dist.names = c("UniFrac",
  "GUniFrac", "WUniFrac", "BC"), grp.name)
}
\arguments{
\item{grp.name}{}
}
\description{
Title
}
