% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{perform_differential_analysis}
\alias{perform_differential_analysis}
\title{Perform differential analysis}
\usage{
perform_differential_analysis(
  data.obj,
  grp.name,
  adj.name = NULL,
  subject = NULL,
  taxa.levels = c("Phylum", "Order", "Class", "Family", "Genus", "Species"),
  method = "perm",
  block.perm = FALSE,
  perm.no = 999,
  norm = "GMPR",
  norm.level = "Species",
  intersect.no = 4,
  ct.min = 2,
  transform = "sqrt",
  winsor = FALSE,
  winsor.qt = 0.97,
  prev = 0.1,
  minp = 0.002,
  medianp = NULL,
  cv = NULL,
  mt.method = "fdr",
  cutoff = 0.15,
  ann = "",
  seed = 123,
  ...
)
}
\arguments{
\item{data.obj}{Data object created by load_data()}

\item{grp.name}{Variable of interest}

\item{adj.name}{List of covariates / variables to adjust for}

\item{subject}{Variable indicating subject}

\item{taxa.levels}{Taxa levels to include in analysis}

\item{method}{Method to use for differential analysis (Default: "perm")}

\item{block.perm}{Perform block permutation (TRUE/FALSE, Default: FALSE)}

\item{perm.no}{Number of permutations to perform (Default: 999)}

\item{norm}{Normalization method (Default: "GMPR")}

\item{norm.level}{Normalization method (Default: "Species")}

\item{intersect.no}{(For GMPR) The minimum number of shared features between sample pair, where the ratio is calculated (Default: 4)}

\item{ct.min}{(For GMPR) The minimum number of counts required to calculate ratio (Default: 2)}

\item{transform}{Type of transformation to perform (Default: "sqrt")}

\item{winsor}{Perform Winsorization of data? (TRUE/FALSE, Default: FALSE)}

\item{winsor.qt}{Winsorization quartile (Default: 0.97)}

\item{prev}{Prevalence cutoff for OTUs (Default: 0.1)}

\item{minp}{Minimum proportion/abundance cutoff (Default: 0.002)}

\item{medianp}{Median proportion/abundance cutoff (Optional, can be used in place of \code{minp}. Default: NULL)}

\item{cv}{Coefficient of variance to use as OTU/ASV cutoff (Optional, can be used in place of \code{minp}. Default: NULL)}

\item{mt.method}{Method to use for multiple testing correction (Default: "fdr")}

\item{cutoff}{Q-value cutoff for significance testing after multiple testing correction}

\item{seed}{Random seed}

\item{...}{Any additional parameters}
}
\value{
A dataframe containing the results of differential abundance testing
}
\description{
Perform differential analysis
}
\examples{
data("Constipation")
diff.obj <- perform_differential_analysis(data.obj, grp.name="Visit",
                                          taxa.levels=c('Phylum', 'Class', 'Order', 'Family', 'Genus', 'Species'),
                                          method='perm', mt.method='fdr',
                                          block.perm=FALSE, norm = 'GMPR')
}
