% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{prev_abund_tables}
\alias{prev_abund_tables}
\title{Prevalence and abundance table generation}
\usage{
prev_abund_tables(data.obj)
}
\arguments{
\item{data.obj}{A data object created by load_data()}
}
\value{
A list of tables(phy.prev, fam.prev, gen.prev, phy.abund, fam.abund, and gen.abund) containing the prevalences and abundances for all Phyla, Families, and Genera in \code{data.obj}
}
\description{
Prevalence and abundance table generation
}
