% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobj_functions.R
\name{subset_data}
\alias{subset_data}
\title{Subset Data object}
\usage{
subset_data(data.obj, samIDs)
}
\arguments{
\item{data.obj}{The data object to subset}

\item{samIDs}{The list of sample IDs to include in the subsetted object}
}
\value{
A subsetted data object containing only the sample IDs in samIDs
}
\description{
Subset Data object
}
