% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobj_functions.R
\name{subset_dist}
\alias{subset_dist}
\title{Subset distance object}
\usage{
subset_dist(dist.obj, samIDs)
}
\arguments{
\item{samIDs}{The list of sample IDs to include in the subsetted object}

\item{data.obj}{The distance object to subset}
}
\value{
A subsetted distance object containing only the sample IDs in samIDs
}
\description{
Subset distance object
}
