% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_plot}
\alias{summary_plot}
\title{Basic summary plots}
\usage{
summary_plot(data.obj, type = "")
}
\arguments{
\item{data.obj}{Data object created by load_data()}

\item{type}{Type of plot to generate, either "abundance", "depth" or "prevalence"}
}
\value{
A ggplot object based on the type specified by \code{type}
}
\description{
Basic summary plots
}
\examples{
summary_plot(data.obj, type="abundance")

}
