% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{visualize_differential_analysis}
\alias{visualize_differential_analysis}
\title{Title}
\usage{
visualize_differential_analysis(data.obj, diff.obj, grp.name = NULL,
  strata = NULL, test = "Nonpara", mt.method = "fdr",
  scale = "sqrt", cutoff = 0.15, taxa.levels = c("Phylum", "Family",
  "Genus"), ord = TRUE, eff.type = "logP", indivplot = TRUE,
  colFnsC = NULL, colFnsF = NULL, subject = NULL, xsize = 10,
  ann = "", hei1 = NULL, wid1 = NULL, hei2 = NULL, wid2 = NULL)
}
\arguments{
\item{wid2}{}
}
\description{
Title
}
