% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{generate_alpha_boxplot}
\alias{generate_alpha_boxplot}
\title{Title}
\usage{
generate_alpha_boxplot(data.obj, alpha.obj = NULL, rarefy = TRUE,
  depth = NULL, iter.no = 5, measures = c("Observed", "Chao1",
  "Shannon", "InvSimpson"), seed = 123, grp.name, strata = NULL,
  gg.cmd = NULL, ann = "", subject = NULL, p.size = 2.5,
  l.size = 0.5, hei = NULL, wid = NULL)
}
\arguments{
\item{wid}{}
}
\description{
Title
}
