% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{perform_alpha_test}
\alias{perform_alpha_test}
\title{Title}
\usage{
perform_alpha_test(data.obj, alpha.obj = NULL, rarefy = TRUE,
  depth = NULL, iter.no = 5, measures = c("Observed", "Chao1",
  "Shannon", "InvSimpson"), model = "lm", formula = NULL,
  grp.name = NULL, adj.name = NULL, subject = NULL, ann = "",
  seed = 123, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
