% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{perform_differential_analysis}
\alias{perform_differential_analysis}
\title{Title}
\usage{
perform_differential_analysis(data.obj, grp.name, adj.name = NULL,
  subject = NULL, taxa.levels = c("Phylum", "Order", "Class", "Family",
  "Genus", "Species"), method = "perm", block.perm = FALSE,
  perm.no = 999, norm = "GMPR", norm.level = "Species",
  intersect.no = 4, ct.min = 2, transform = "sqrt", winsor = FALSE,
  winsor.qt = 0.97, prev = 0.1, minp = 0.002, medianp = NULL,
  cv = NULL, mt.method = "fdr", cutoff = 0.15, ann = "",
  seed = 123, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
