% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{perform_mirkat_test}
\alias{perform_mirkat_test}
\title{Title}
\usage{
perform_mirkat_test(data.obj, dist.obj, dist.names = c("UniFrac",
  "GUniFrac", "WUniFrac", "BC"), grp.name = NULL, adj.name = NULL,
  pairwise = F, ann = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
