% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{perform_permanova_test}
\alias{perform_permanova_test}
\title{Title}
\usage{
perform_permanova_test(data.obj, dist.obj, dist.names = c("UniFrac",
  "GUniFrac", "WUniFrac", "BC"), PermanovaG.dist = c("UniFrac",
  "GUniFrac", "WUniFrac", "BC"), formula = NULL, grp.name = NULL,
  adj.name = NULL, pairwise = F, block.perm = F, strata = NULL,
  ann = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
